refactor(maven): use maven module in monolithic app
