+++ base-info/src/main/java/com/smartagilify/baseinfo/dtos/BaseInfoDTO.java
