+++ base-info/src/main/java/com/smartagilify/baseinfo/dtos/BaseInfoDetailDTO.java
