+++ base-info/src/main/java/com/smartagilify/baseinfo/entities/BaseInfoDetail.java
