+++ base-info/src/main/java/com/smartagilify/baseinfo/exception/BaseInfoNotFoundException.java
