+++ base-info/src/main/java/com/smartagilify/baseinfo/exception/ExceptionMessageEnum.java
