+++ base-info/src/main/java/com/smartagilify/baseinfo/repositories/BaseInfoDetailRepository.java
@@ -1,6 +1,5 @@
 package com.smartagilify.baseinfo.repositories;
 
-import com.smartagilify.baseinfo.dtos.BaseInfoDetailDTO;
 import com.smartagilify.baseinfo.entities.BaseInfoDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
