+++ base-info/src/main/java/com/smartagilify/baseinfo/repositories/BaseInfoRepository.java
@@ -1,6 +1,5 @@
 package com.smartagilify.baseinfo.repositories;
 
-import com.smartagilify.baseinfo.dtos.BaseInfoDTO;
 import com.smartagilify.baseinfo.entities.BaseInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
