+++ base-info/src/main/resources/application.yml
