+++ project-management/src/main/java/com/smartagilify/projectmanagement/ProjectManagementApplication.java
@@ -1,14 +1,14 @@
-package com.smartagilify;
+package com.smartagilify.projectmanagement;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 
 @SpringBootApplication(exclude = {SecurityAutoConfiguration.class })
-public class SmartAgilifyApplication {
+public class ProjectManagementApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(SmartAgilifyApplication.class, args);
+        SpringApplication.run(ProjectManagementApplication.class, args);
     }
 
 }
