+++ project-management/src/main/java/com/smartagilify/projectmanagement/controllers/TestController.java
@@ -1,0 +1,12 @@
+package com.smartagilify.projectmanagement.controllers;
+
+import com.smartagilify.baseinfo.entities.BaseInfo;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class TestController {
+
+    public void test() {
+        BaseInfo bi = new BaseInfo();//FIXME remove it
+    }
+}
